Move Javadoc base extension to documentation kit
