+++ gradle/plugins/documentation-kit-plugins/javadoc/src/main/java/dev/gradleplugins/dockit/javadoc/JavadocSourcePathsOption.java
@@ -13,9 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nokeebuild.javadoc;
+package dev.gradleplugins.dockit.javadoc;
 
-import com.google.common.collect.Iterables;
 import org.gradle.api.Action;
 import org.gradle.api.Project;
 import org.gradle.api.Transformer;
@@ -25,13 +24,12 @@ import org.gradle.api.provider.ListProperty;
 import org.gradle.api.tasks.javadoc.Javadoc;
 
 import java.io.File;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 import static java.util.Arrays.asList;
 import static java.util.Collections.emptyList;
-import static java.util.stream.Collectors.toList;
 import static java.util.stream.StreamSupport.stream;
 
 /**
@@ -55,9 +53,9 @@ public final class JavadocSourcePathsOption implements Action<Javadoc> {
 		args.value(sourcePaths.getElements().map(asSourcePathFlags()).orElse(emptyList())).disallowChanges();
 	}
 
-	private static Transformer<Iterable<String>, Iterable<FileSystemLocation>> asSourcePathFlags() {
+	private static Transformer<Iterable<String>, Collection<FileSystemLocation>> asSourcePathFlags() {
 		return values -> {
-			if (Iterables.isEmpty(values)) {
+			if (values.isEmpty()) {
 				return Collections.emptyList();
 			} else {
 				return asList("-sourcepath", stream(values.spliterator(), false).map(it -> it.getAsFile().getAbsolutePath()).collect(Collectors.joining(File.pathSeparator)));
