--- gradle/plugins/nokeebuild-plugins/jvm-basics/src/main/java/nokeebuild/javadoc/JavadocBasePlugin.java
@@ -1,40 +1,0 @@
-/*
- * Copyright 2022 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package nokeebuild.javadoc;
-
-import org.gradle.api.Plugin;
-import org.gradle.api.Project;
-import org.gradle.api.tasks.TaskCollection;
-import org.gradle.api.tasks.javadoc.Javadoc;
-
-import javax.inject.Inject;
-
-class JavadocBasePlugin implements Plugin<Project> {
-	@Inject
-	public JavadocBasePlugin() {}
-
-	@Override
-	public void apply(Project project) {
-		final TaskCollection<Javadoc> javadocTasks = project.getTasks().withType(Javadoc.class);
-		javadocTasks.configureEach(new JavadocLinksOption(project));
-		javadocTasks.configureEach(new JavadocSourcePathsOption(project));
-		javadocTasks.configureEach(new JavadocTitleOption(project));
-		javadocTasks.configureEach(new JavadocExcludeOption(project));
-		javadocTasks.configureEach(new JavadocSubpackagesOption(project));
-		javadocTasks.configureEach(new JavadocAdditionalArgsOption(project));
-		javadocTasks.configureEach(new JavadocSourcesOption(project));
-	}
-}
